colosseum-agent-hackathon/agent/pyproject.toml
[project]
name = "solshield-agent"
version = "0.1.0"
description = "AI-Powered Liquidation Prevention Agent for Solana"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "solana>=0.34.0",
    "solders>=0.21.0",
    "anchorpy>=0.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "apscheduler>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
